Some partial changes made the last week

+++ specs/trunk/zaptel.spec
@@ -2,16 +2,17 @@
 # if none was specified with "--define 'kernel <uname -r>'"
 %{!?kernel: %{expand: %%define kernel %(uname -r)}}
 
-%define kversion %(echo %{kernel} | sed -e s/smp// -)
+%define kversion %(echo %{kernel} | sed -e s/xen// -)
 %define krelver  %(echo %{kversion} | tr -s '-' '_')
-%if %(echo %{kernel} | grep -c smp)
-        %{expand:%%define ksmp -smp}
+
+%if %(echo %{kernel} | grep -c xen)
+        %{expand:%%define kxensufix -xen}
 %endif
 
 Summary: Telephony interface support
 Name: zaptel
 Version: 1.4.5.1
-Release: 21%{?lptver}
+Release: 25
 License: GPL
 Group: System Environment/Libraries
 URL: http://www.asterisk.org/
@@ -21,11 +22,11 @@ Patch0: ystdm8xx-zaptel-1.4.5.patch
 #Patch1: zaptel-sysconfig-yeastar.patch
 Patch2: zaptel-init.patch
 Patch3: zaptel-zma8xx-1.4.5.1.patch
-#Patch4: zaptel-oslec-1.4.5.1.patch
-Patch5: zaptel-octasic-1.4.5.1.patch
+Patch4: zaptel-oslec-1.4.5.1.patch
+#Patch5: zaptel-octasic-1.4.5.1.patch
 Prereq: kernel-module-zaptel, kernel-module-zaptel-xen
 BuildRoot: %{_tmppath}/%{name}-%{version}-%{release}-root
-BuildRequires: kernel%{?ksmp}-devel = %{kversion}
+BuildRequires: kernel%{?kxensufix}-devel = %{kversion}
 BuildRequires: newt-devel, libusb-devel, MAKEDEV
 #BuildRequires: newt-devel, MAKEDEV
 
@@ -46,26 +47,26 @@ Requires: %{name} = %{version}
 This package contains the header files needed to compile applications that
 will use Zaptel, such as Asterisk.
 
-%package -n kernel%{?ksmp}-module-zaptel
+%package -n kernel-module-zaptel%{?kxensufix}
 Summary: Kernel modules required for some hardware to operate with Zaptel
 Release: %{release}_%{krelver}
 Group: System Environment/Kernel
-Requires: kernel%{?ksmp} = %{kversion}, /sbin/depmod
+Requires: kernel%{?kxensufix} = %{kversion}, /sbin/depmod
 Provides: kernel-modules
-%{?ksmp:Provides: kernel-module-zaptel = %{version}-%{release}_%{krelver}}
+%{?kxensufix:Provides: kernel-module-zaptel%{?kxensufix} = %{version}-%{release}_%{krelver}}
 #%kmdl_dependencies
 
-%description -n kernel%{?ksmp}-module-zaptel
+%description -n kernel-module-zaptel%{?kxensufix}
 This package contains the zaptel kernel modules for the Linux kernel package :
-%{kversion} (%{_target_cpu}%{?ksmp:, SMP}).
+%{kversion} (%{_target_cpu}%{?kxensufix:, XEN}).
 
-%package octasic
-Summary: Support for the Octasic echo canceller in Elastix
-Group: System Environment/Kernel
-Requires: %{name} = %{version}
+#%package octasic
+#Summary: Support for the Octasic echo canceller in Elastix
+#Group: System Environment/Kernel
+#Requires: %{name} = %{version}
 
-%description octasic
-Support for the Octasic echo canceller in Elastix 
+#%description octasic
+#Support for the Octasic echo canceller in Elastix 
 
 %prep
 %setup
@@ -73,8 +74,8 @@ Support for the Octasic echo canceller in Elastix
 #%patch1 -p0 
 %patch2 -p0 
 %patch3 -p1
-#%patch4 -p1
-%patch5 -p1
+%patch4 -p1
+#%patch5 -p1
 # Fix lib vs. lib64
 %{__perl} -pi -e 's|/usr/lib|%{_libdir}|g' Makefile
 # Force mknod calls to never happen
@@ -141,29 +142,6 @@ ${MAKEDEV} \
 %{__install} -Dp -m0755 zaptel.conf.sample \
     %{buildroot}%{_sysconfdir}/zaptel.conf
 
-# Install Octasic support
-cd octvqe
-%{__make} \
-    KVERS="%{kernel}" \
-    INSTALL_PREFIX="%{buildroot}" \
-    DESTDIR="%{buildroot}" \
-    ROOT_PREFIX="%{buildroot}"
-
-%{__mkdir_p} %{buildroot}/lib/modules/%{kernel}/misc/octvqe 
-%{__install} -Dp -m0755 octvqe.ko %{buildroot}/lib/modules/%{kernel}/misc/octvqe/
-%{__install} -Dp -m0755 octvqe.o %{buildroot}/lib/modules/%{kernel}/misc/octvqe/
-
-#%{__install} -Dp -m0755 octvqed.init %{buildroot}%{_sysconfdir}/init.d/octvqed
-#%{__install} -Dp -m0755 octvqed.conf %{buildroot}%{_sysconfdir}/
-
-# Instalo el daemon octvqed, pero depende de la plataforma asi que a lo mejor
-# puedo instalar 2 versiones. Una para AMD y otra para INTEL y en el post
-# recien elijo cual instalar dependiendo de la arquitectura del host
-#%{__mkdir_p} %{buildroot}/usr/sbin
-#%{__install} -Dp -m0755 octvqed %{buildroot}/usr/sbin/
-# Falta el register, pero esto solo es poner el binario de 32 bits
-#%{__install} -Dp -m0755 register32 %{buildroot}/usr/sbin/
-
 %clean
 %{__rm} -rf %{buildroot}
 
@@ -176,7 +154,12 @@ cd octvqe
 mkdir -p /var/lib/digium/licenses
 
 # To maintain compatibility with the older genzaptelconf location in Elastix
-ln -s /usr/sbin/genzaptelconf /usr/local/sbin/genzaptelconf
+if [ -f "/usr/local/sbin/genzaptelconf" ]; then
+    mv --force /usr/local/sbin/genzaptelconf /usr/local/sbin/genzaptelconf.old
+    ln -s /usr/sbin/genzaptelconf /usr/local/sbin/genzaptelconf
+else
+    ln -s /usr/sbin/genzaptelconf /usr/local/sbin/genzaptelconf
+fi
 
 %postun 
 /sbin/ldconfig
@@ -184,12 +167,11 @@ ln -s /usr/sbin/genzaptelconf /usr/local/sbin/genzaptelconf
 # activate the original zaptel udev rules in /etc/udev/rules.d/50-udev.rules
 # perl -pi -e's,#KERNEL\=\=\"zap,KERNEL\=\=\"zap,g' /etc/udev/rules.d/50-udev.rules
 
-
-%post -n kernel%{?ksmp}-module-zaptel
+%post -n kernel-module-zaptel%{?kxensufix}
 /sbin/depmod -a -F /boot/System.map-%{kernel} %{kernel} &>/dev/null || :
 
-%postun -n kernel%{?ksmp}-module-zaptel
-/sbiin/depmod -a -F /boot/System.map-%{kernel} %{kernel} &>/dev/null || :
+%postun -n kernel-module-zaptel%{?kxensufix}
+/sbin/depmod -a -F /boot/System.map-%{kernel} %{kernel} &>/dev/null || :
 
 %files -f device.list
 %defattr(-, root, root, 0755)
@@ -210,49 +192,42 @@ ln -s /usr/sbin/genzaptelconf /usr/local/sbin/genzaptelconf
 %{_libdir}/*.so.*
 %{_libdir}/*.a
 %{_mandir}/man8/*
-#%exclude /usr/sbin/octvqed
-#%exclude /usr/sbin/register32
 
 %files devel
 %defattr(-, root, root, 0755)
 %{_includedir}/zaptel/*.h
 %{_libdir}/*.so
 
-%files -n kernel%{?ksmp}-module-zaptel
+%files -n kernel-module-zaptel%{?kxensufix}
 %defattr(-, root, root, 0755)
 /lib/modules
-#/lib
-%exclude /lib/modules/%{kernel}/misc/octvqe/octvqe.ko
-%exclude /lib/modules/%{kernel}/misc/octvqe/octvqe.o
-
-%files octasic
-%defattr(-, root, root, 0755)
-/lib/modules/%{kernel}/misc/octvqe/octvqe.ko
-/lib/modules/%{kernel}/misc/octvqe/octvqe.o
-#%{_sysconfdir}/init.d/octvqed
-#%{_sysconfdir}/octvqed.conf
-# Falta el daemon y el register32
-#/usr/sbin/octvqed
-#/usr/sbin/register32
 
 %changelog
+* Thu Oct 20 2007 Edgar Landivar <elandivar@palosanto.com>1.4.5.1-25
+- Better XEN support when building the RPM.
+- Fixed minor typo error in the postun section related with the depmod invocation.
+
+* Thu Oct 18 2007 Edgar Landivar <elandivar@palosanto.com>1.4.5.1-23
+- Added logic to check if there is a previous genzaptelconf version.
+- Octasic support removed. OSLEC is the official Elastix echo canceller now.
+
 * Tue Oct 11 2007 Edgar Landivar <elandivar@palosanto.com>1.4.5.1-21
-- Some changes regarding Octasic support. Some files will belong to a new 
+- Some changes regarding Octasic support. Some files will belong to a new
   octasic package. Just the octasic modules remain in the zaptel package.
 
 * Tue Oct  9 2007 Edgar Landivar <elandivar@palosanto.com>1.4.5.1-20
-- Symbolic Link from /usr/local/sbin/genzaptelconf to /usr/sbin/genzaptelconf 
-  to maintain compatibility with the older genzaptelconf location in Elastix
+- Symbolic Link from /usr/local/sbin/genzaptelconf to /usr/sbin/genzaptelconf
+  to maintain compatibility with the older genzaptelconf location in Elastix.
 
 * Fri Sep 14 2007 Edgar Landivar <elandivar@palosanto.com>1.4.5.1-12
-- Octasic echo canceller support added
-- OSLEC temporally removed because it does not work against this zaptel version
+- Octasic echo canceller support added.
+- OSLEC temporally removed because it does not work against this zaptel version.
 
 * Tue Sep 11 2007 Edgar Landivar <elandivar@palosanto.com>1.4.5.1-6
-- Zapmicro ZMA800 patch added
+- Zapmicro ZMA800 patch added.
 
 * Tue Sep 11 2007 Edgar Landivar <elandivar@palosanto.com>1.4.5.1-4
-- OSLEC echo canceller support added 
+- OSLEC echo canceller support added.
 
 * Wed Sep 5 2007 Edgar Landivar <elandivar@palosanto.com>1.4.5.1-2
 - Removing the libusb-devel requirement until the next version, when that
